refactor(codegen): split extkey and symbol into separate modules
- Move extkey.py to its own module for better organization
- Create symbol.py module for special symbol handling
- Update gen.py to import from new modules
- Remove ext_keys and Prefix from lib.py
